experimental feature: allow mismatches in barcodes
algorithm is as follows: Each generated unique barcode is matched against the
beginning of the read. The found number of mismatches is calculated back to
the length of the full barcode (e.g. full barcode = ACTGACTG, we match GATCTG
with 2 mismatches, then back-calculated mismatches are 2 / 6 * 8 = 2.67.
The best-matching barcode is accepted if the back-calculated number of mismatches
for this barcode is below the maximum allowed number of mismatches